Refactor basic data form
