+++ src/views/registration/FormContainer.js
@@ -3,16 +3,20 @@ import {
   KeyboardAvoidingView,
   ScrollView,
   StyleSheet,
+  Text,
   View,
 } from "react-native";
 import PropTypes from "prop-types";
-import { Colors } from "../../constants/styles";
+import { Colors, Typography } from "../../constants/styles";
 
-const FormContainer = ({ children }) => {
+const FormContainer = ({ title, children }) => {
   return (
     <KeyboardAvoidingView style={styles.backgroundContainer}>
       <ScrollView>
-        <View style={styles.container}>{children}</View>
+        <View style={styles.container}>
+          {title && <Text style={styles.titleText}>{title}</Text>}
+          {children}
+        </View>
       </ScrollView>
     </KeyboardAvoidingView>
   );
@@ -29,10 +33,22 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 50,
     paddingTop: 22,
   },
+  titleText: {
+    marginLeft: 30,
+    marginBottom: 20,
+    color: Colors.PURPLE,
+    fontSize: Typography.FONT_SIZE_16,
+    fontFamily: Typography.FONT_FAMILY_BOLD,
+  },
 });
+
+FormContainer.defaultProps = {
+  title: "",
+};
 
 FormContainer.propTypes = {
   children: PropTypes.node.isRequired,
+  title: PropTypes.string,
 };
 
 export default FormContainer;
