+++ src/views/registration/PhysicalExamination.js
@@ -685,11 +685,9 @@ const PhysicalExamination = ({ route, navigation }) => {
                   numberOfLines={2}
                 />
                 <AppButton
-                  buttonType="solid"
                   icon="next_btn"
                   onPress={handleSubmit}
                   disabled={!isValid || isSubmitting || isNextButtonDisabled}
-                  loading={isSubmitting}
                 />
               </>
             )}
